Fixes panic when reading non-conformant tests
Introduces flag `--reorder_all` which must be used in case of a
non-conformant test that prints annotations before test result.

This is needed in cases when you have non-conformant test output such
as in https://github.com/bats-coore/issues/187.  Ideally the test
harness should capture output and print following the test, but
in that particular case it is not easy to do.

Fixes #6.